Move more components and functions out
