+++ client/Layout/CenterColumn.tsx
@@ -1,15 +1,13 @@
 import * as React from "react";
 import { TrackerViewModel } from "../TrackerViewModel";
 import { useSubscription } from "../Combatant/linkComponentToObservables";
-import {
-  StatBlockEditor,
-  StatBlockEditorProps
-} from "../StatBlockEditor/StatBlockEditor";
-import { SpellEditor, SpellEditorProps } from "../StatBlockEditor/SpellEditor";
+import { StatBlockEditor } from "../StatBlockEditor/StatBlockEditor";
+import { SpellEditor } from "../StatBlockEditor/SpellEditor";
 import { PendingPrompts } from "../Prompts/PendingPrompts";
 import { CombatFooter } from "../CombatFooter/CombatFooter";
 import { InitiativeListHost } from "./InitiativeListHost";
 import { useVerticalResizerDrop } from "./useVerticalResizerDrop";
+import { centerColumnView } from "./centerColumnView";
 
 export function CenterColumn(props: { tracker: TrackerViewModel }) {
   const statblockEditorProps = useSubscription(
@@ -40,54 +38,4 @@ export function CenterColumn(props: { tracker: TrackerViewModel }) {
       />
     </div>
   );
-}
-export function centerColumnView(
-  statBlockEditorProps: StatBlockEditorProps,
-  spellEditorProps: SpellEditorProps
-) {
-  if (statBlockEditorProps !== null) {
-    return "statblockeditor";
-  }
-  if (spellEditorProps !== null) {
-    return "spelleditor";
-  }
-  return "combat";
-}
-export function interfacePriorityClass(
-  centerColumnView: string,
-  librariesVisible: boolean,
-  hasPrompt: boolean,
-  isACombatantSelected: boolean,
-  encounterState: "active" | "inactive"
-) {
-  if (
-    centerColumnView === "statblockeditor" ||
-    centerColumnView === "spelleditor"
-  ) {
-    if (librariesVisible) {
-      return "show-center-left-right";
-    }
-    return "show-center-right-left";
-  }
-
-  if (librariesVisible) {
-    return "show-left-center-right";
-  }
-
-  if (hasPrompt) {
-    if (isACombatantSelected) {
-      return "show-center-right-left";
-    }
-    return "show-center-left-right";
-  }
-
-  if (isACombatantSelected) {
-    return "show-right-center-left";
-  }
-
-  if (encounterState == "active") {
-    return "show-center-left-right";
-  }
-
-  return "show-center-right-left";
 }
